Moves frontend files around and updates to ES2015 modules.
