+++ public/index.html
@@ -4,6 +4,6 @@
 <body>
   <canvas id="garden" width="400" height="300"></canvas>
   <canvas id="chatter" width="400" height="300"></canvas>
-  <script type="text/javascript" src='bundle.js'></script>
+  <script type="module" src="peeps.js"></script>
 </body>
 </html>
